Move: Moved account creation logic into its appropriate place inside the models directory
