+++ backend/models/user.js
@@ -17,4 +17,15 @@ async function getUserById(id) {
   return user[0][0];
 }
 
-module.exports = { getUserByEmail, getUserById };
+async function createAccount(userInput) {
+  try {
+    const query = `insert into users (email,full_name,phone, pass) values(?, ?, ?, ?);`;
+
+    const results = await pool.query(query, userInput);
+    return { success: true };
+  } catch (error) {
+    return { success: false, error: error };
+  }
+}
+
+module.exports = { getUserByEmail, getUserById, createAccount };
